requirements.txt
streamlit
langchain~=0.3.14
langchain-community
langchain-aws
amazon-bedrock
transformers
requests~=2.32.3
boto3~=1.35.92
pandas
matplotlib
reportlab
python-telegram-bot
beautifulsoup4
selenium~=4.21.0
textblob
python-dotenv
pytest